Implementing Clients DTO (Data Transfer Object) and automapper for safer web api work

+++ NextYearUniversityExam/Exam2/ImprovedExamWithApi/Controllers/ExamController.cs
@@ -1,7 +1,9 @@
-﻿using Domain.Models;
+﻿using AutoMapper;
+using Domain.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Services.Services;
+using Services.DTOs.ClientDTOs;
 
 namespace ImprovedExamWithApi.Controllers
 {
@@ -10,10 +12,12 @@ namespace ImprovedExamWithApi.Controllers
     public class ExamController : ControllerBase
     {
         private readonly ExamService _service;
+        private readonly IMapper _mapper;
 
-        public ExamController(ExamService service)
+        public ExamController(ExamService service, IMapper mapper)
         {
-            _service = service; 
+            _service = service;
+            _mapper = mapper;
         }
 
         [HttpGet]
@@ -45,29 +49,54 @@ namespace ImprovedExamWithApi.Controllers
             return Ok(group);
 
         }
-
-        [HttpPost("Add-Client")]
-        public async Task AddClient(Client obj)
+        [HttpGet("Get-All-Clients")]
+        public async Task<ActionResult<IEnumerable<ClientReadDto>>> GetAllClients()
         {
-            await _service.AddClient(obj);
+            var clients = await _service.GetClients();
+            var clientsDto = _mapper.Map<IEnumerable<ClientReadDto>>(clients);
+            return Ok(clientsDto);
+        }
+
+        [HttpPost("Add-Client", Name = "AddClient")]
+        public async Task<ActionResult<ClientCreateDto>> AddClient(ClientCreateDto obj)
+        {
+            var ClientEntity = _mapper.Map<Client>(obj);
+
+            try
+            {
+                await _service.AddClient(ClientEntity);
+                var clientReadDto = _mapper.Map<ClientReadDto>(ClientEntity);
+
+                return CreatedAtRoute("AddClient", new { Id = clientReadDto.Id }, clientReadDto);
+            }
+            catch (Exception ex) { return BadRequest(ex.Message); }
+
+          
+
         }
 
         [HttpPut("Update-Client")]
-        public async Task <IActionResult> UpdateClient(Client obj)
+        public async Task <IActionResult> UpdateClient(int id, ClientUpdateDto obj)
         {
+            var existingClient = await _service.GetClientById(id);
+            if (existingClient == null) return NotFound();
+
+            _mapper.Map(obj, existingClient); // automapper do all hardwork for us 
+
             try
             {
-                await _service.UpdateClient(obj);
-                return Ok();
+                await _service.UpdateClient(existingClient);
+                return NoContent();
             }catch(Exception ex) { return BadRequest(ex.Message); }
             
         }
         [HttpDelete("Delete-Client")]
         public async Task<IActionResult> DeleteClient(int id)
         {
+            var delete = await _service.GetClientById(id);
+            if (delete == null) return NotFound("Client is not found");
             try
             {
-                var delete = await _service.GetClientById(id);
                 await _service.RemoveClient(delete);
                 return Ok();
             }catch(Exception ex) { return BadRequest(ex.Message); }
